move CountedBitSet back to translog package
